prepare for quotients hints and verifier hints
